Added game timer

+++ Pirate Game/Assets/Scripts/Save/SaveData.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [System.Serializable]
 public class SaveData
 {
-    public int maxClockTime;
+    public float maxClockTime;
     public Vector3 playerStartPosition;
     public float defaultEnemySpawnTime;
